[#91] refactor: 주문 생성 리팩토링 및 예외처리

+++ src/main/java/com/sparta/spangeats/domain/order/dto/OrderResponse.java
@@ -20,7 +20,7 @@ public record OrderResponse(
         DeliveryType deliveryType,
         LocalDateTime createdAt,
         LocalDateTime updatedAt,
-        Integer totalPrice
+        Long totalPrice
 ) {
     public static OrderResponse from(Order order) {
         return OrderResponse.builder()
