+++ src/main/java/com/sparta/spangeats/domain/order/dto/OrderSaveRequest.java
@@ -17,6 +17,6 @@ public record OrderSaveRequest(
         PaymentType paymentType,
         DeliveryType deliveryType,
         List<MenuOrders> orders,
-        Integer totalPrice
+        Long totalPrice
 ) {
 }
